fixed formatting, changed to table driven test
